Refactor download notebook
